Add .cargo to gitignore
To use a specific version of MLIR, it is useful to set the corresponding environment variable in `.cargo/config.toml`. In order to use this while working in melior, I think it makes sense to gitignore the `.cargo` directory. This can be reverted if the project ever needs to set it to something specific.